pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "Backend сервис для плагина Sber IDE (каркас)."
authors = [{ name = "Agent Service" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.30.1",
    "pydantic==2.8.2",
    "pydantic-settings==2.4.0",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "boto3==1.35.99",
    "langchain==0.3.19",
    "langchain-core==0.3.40",
    "langchain-community==0.3.18",
    "langgraph==0.2.53",
    "chromadb==0.5.4",
    "gigachat==0.1.21",
]

[project.scripts]
agent-service = "app.main:main"
agent-service-worker = "app.worker:main"
agent-service-tool-host = "app.tool_host:main"
claude-code-adapter = "claude_code_adapter_app.main:main"
agent-service-claude-code-adapter = "claude_code_adapter_app.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
